+++ rpc/types_contract.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"math/big"
 	"strconv"
 	"strings"
 
@@ -19,6 +20,7 @@ type NumAsHex string
 type U64 string
 
 // ToUint64 converts the U64 type to a uint64.
+// If the value is greater than max uint64, returns an error.
 func (u U64) ToUint64() (uint64, error) {
 	hexStr := strings.TrimPrefix(string(u), "0x")
 
@@ -32,6 +34,27 @@ func (u U64) ToUint64() (uint64, error) {
 
 // 128 bit unsigned integers, represented by hex string of length at most 32
 type U128 string
+
+// ToBigInt converts the U128 type to a *big.Int.
+// If the value is greater than max uint128, returns an error.
+//
+//nolint:mnd // 16 means hex base
+func (u U128) ToBigInt() (*big.Int, error) {
+	hexStr := strings.TrimPrefix(string(u), "0x")
+
+	result, ok := new(big.Int).SetString(hexStr, 16)
+	if !ok {
+		return nil, fmt.Errorf("failed to parse hex string: %v", hexStr)
+	}
+
+	maxUint128, _ := new(big.Int).SetString("ffffffffffffffffffffffffffffffff", 16)
+
+	if result.Cmp(maxUint128) > 0 {
+		return nil, fmt.Errorf("value is greater than max uint128: %v", u)
+	}
+
+	return result, nil
+}
 
 type ClassOutput interface{}