+++ utils/transactions_test.go
@@ -12,10 +12,14 @@ import (
 )
 
 func TestResBoundsMapToOverallFee(t *testing.T) {
+	t.Parallel()
+
+	zeroTip := rpc.U64("0x0")
 	tests := []struct {
 		name        string
 		resBounds   rpc.ResourceBoundsMapping
 		multiplier  float64
+		tip         rpc.U64
 		expectedRes string
 		expectedErr string
 	}{
@@ -36,6 +40,7 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier: 1.0,
+			tip:        zeroTip,
 			// Expected: (100*100) + (100*100) + (100*100) = 10000 + 10000 + 10000 = 30000
 			expectedRes: "30000",
 		},
@@ -56,6 +61,7 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier:  1.5,
+			tip:         zeroTip,
 			expectedRes: "0",
 		},
 		{
@@ -75,6 +81,7 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier: 1.5,
+			tip:        zeroTip,
 			// Expected: ((100*100) + (100*100) + (100*100)) * 1.5 = 30000 * 1.5 = 45000
 			expectedRes: "45000",
 		},
@@ -95,7 +102,8 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier:  -1.0,
-			expectedErr: "multiplier cannot be negative",
+			tip:         zeroTip,
+			expectedErr: "multiplier must be greater than 0",
 		},
 		{
 			name: "Multiplier less than 1",
@@ -114,6 +122,7 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier: 0.5,
+			tip:        zeroTip,
 			// Expected: ((100*100) + (100*100) + (100*100)) * 0.5 = 30000 * 0.5 = 15000
 			expectedRes: "15000",
 		},
@@ -134,6 +143,7 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier:  1.5,
+			tip:         zeroTip,
 			expectedRes: "45000000000000000000000000000000000000000000",
 		},
 		{
@@ -144,6 +154,8 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 					MaxPricePerUnit: "0xa",          // 10
 				},
 			},
+			multiplier:  1.0,
+			tip:         zeroTip,
 			expectedErr: "invalid resource bounds: 'invalidValue' is not a valid big.Int",
 		},
 		{
@@ -163,6 +175,7 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier:  1.0,
+			tip:         zeroTip,
 			expectedErr: "invalid resource bounds: '' is not a valid big.Int",
 		},
 		{
@@ -182,6 +195,7 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier:  1.0,
+			tip:         zeroTip,
 			expectedErr: "can't fit in felt: 0x64000000000000000000000000000000000000000000000000000000000000000004dbc",
 		},
 		{
@@ -193,13 +207,38 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 				},
 			},
 			multiplier:  1.0,
+			tip:         zeroTip,
 			expectedErr: "resource bounds cannot be negative, got '-0x64'",
+		},
+		{
+			name: "Real values",
+			resBounds: rpc.ResourceBoundsMapping{
+				L1Gas: rpc.ResourceBounds{
+					MaxAmount:       "0x0",
+					MaxPricePerUnit: "0x1925a36320fc",
+				},
+				L1DataGas: rpc.ResourceBounds{
+					MaxAmount:       "0x80",   // 128
+					MaxPricePerUnit: "0x6c01", // 27649
+				},
+				L2Gas: rpc.ResourceBounds{
+					MaxAmount:       "0xc25b1",    // 796081
+					MaxPricePerUnit: "0xb2d05e00", // 3000000000
+				},
+			},
+			multiplier: 1.0,
+			tip:        rpc.U64("0x1000000"), // 16777216
+			// Expected: 0 + (128*27649) + ((3000000000+16777216)*796081) =
+			// 0 + 3539072 + 2401599022890496 = 2401599026429568
+			expectedRes: "2401599026429568", // 0x8883dd8dcfe80
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := ResBoundsMapToOverallFee(&tt.resBounds, tt.multiplier)
+			t.Parallel()
+
+			got, err := ResBoundsMapToOverallFee(&tt.resBounds, tt.multiplier, tt.tip)
 
 			if tt.expectedErr != "" {
 				require.Error(t, err)
@@ -221,13 +260,16 @@ func TestResBoundsMapToOverallFee(t *testing.T) {
 	}
 }
 
-//nolint:dupl // The tests are similar, but they are testing different things.
+//nolint:dupl,tparallel // The tests are similar, but they are testing different things. // Run sequentially to avoid race conditions with the `tests` variable.
 func TestFeeEstToResBoundsMap(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
 		name          string
 		feeEstimation rpc.FeeEstimation
 		multiplier    float64
 		expected      rpc.ResourceBoundsMapping
+		feeLimit      FeeLimits // Only used in the `CustomFeeEstToResBoundsMap` test.
 	}{
 		{
 			name: "Basic calculation with multiplier 1.0",
@@ -283,8 +325,8 @@ func TestFeeEstToResBoundsMap(t *testing.T) {
 					), // invalid uint128
 					L2GasConsumed: internalUtils.TestHexToFelt(
 						t,
-						"0x123456789abcdef0123456789abcdabcdabcd",
-					), // invalid uint64
+						"0x2faf0800",
+					), // valid uint64, within L2 gas amount limit
 				},
 			},
 			multiplier: 0.5,
@@ -302,11 +344,11 @@ func TestFeeEstToResBoundsMap(t *testing.T) {
 					MaxPricePerUnit: "0x517ff0e961da52cb43bc34ad259e168a",
 				},
 				L2Gas: rpc.ResourceBounds{
-					// As these inputs overflow, the multiplier will be applied to the max values
-					// 18446744073709551615 * 0.5 ~= 9223372036854775807
-					MaxAmount: "0x7fffffffffffffff",
+					// 800_000_000 * 0.5 ~= 400_000_000
+					MaxAmount: "0x17d78400",
+					// As the output overflows, the max value is used
 					// 340282366920938463463374607431768211455 * 0.5 ~= 170141183460469231731687303715884105727
-					MaxPricePerUnit: "0x7fffffffffffffffffffffffffffffff",
+					MaxPricePerUnit: rpc.U128(maxUint128),
 				},
 			},
 		},
@@ -384,11 +426,14 @@ func TestFeeEstToResBoundsMap(t *testing.T) {
 				},
 				L2Gas: rpc.ResourceBounds{
 					// 71737338064426034 * 1.7 ~= 121953474709524254
-					MaxAmount: "0x1b1440a032f8f1e",
+					// The result is bigger than the max L2 gas amount limit, so the function
+					// should return the max L2 gas amount instead
+					MaxAmount: rpc.U64(maxL2GasAmount),
 					// 5907679981266292691599931071900621 * 1.7 ~= 10043055968152697313366189329472992
 					MaxPricePerUnit: "0x1ef293003a41145dddddddddddde0",
 				},
 			},
+			feeLimit: starknetLimits, // For the CustomFeeEstToResBoundsMap test.
 		},
 		{
 			name: "Zero values",
@@ -459,17 +504,19 @@ func TestFeeEstToResBoundsMap(t *testing.T) {
 				L1DataGas: rpc.ResourceBounds{
 					// 12406075901516675723 * 1.7 ~= 21090329032578348178
 					// This result is too large to fit in a uint64, so the function returns the max uint64 value
-					MaxAmount: rpc.U64(fmt.Sprintf("%#x", maxUint64)),
+					MaxAmount: rpc.U64(maxUint64),
 					// 216663551256725667606984177334664047893 * 1.7 ~= 368328037136433625310078573378144594674
 					// This result is too large to fit in a uint128, so the function returns the max uint128 value
 					MaxPricePerUnit: rpc.U128(maxUint128),
 				},
+				// The inputs overflow, so the output should be the max values
 				L2Gas: rpc.ResourceBounds{
-					// The inputs overflow, so should the output
-					MaxAmount:       rpc.U64(fmt.Sprintf("%#x", maxUint64)),
+					// Default max L2 gas amount limit
+					MaxAmount:       rpc.U64(maxL2GasAmount),
 					MaxPricePerUnit: rpc.U128(maxUint128),
 				},
 			},
+			feeLimit: starknetLimits, // For the CustomFeeEstToResBoundsMap test.
 		},
 		{
 			name: "Negative multiplier",
@@ -520,27 +567,148 @@ func TestFeeEstToResBoundsMap(t *testing.T) {
 		},
 	}
 
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := FeeEstToResBoundsMap(tt.feeEstimation, tt.multiplier)
+	t.Run("Test FeeEstToResBoundsMap", func(t *testing.T) {
+		for _, tt := range tests {
+			t.Run(tt.name, func(t *testing.T) {
+				got := FeeEstToResBoundsMap(tt.feeEstimation, tt.multiplier)
 
-			// Compare each field individually for better error messages
-			assert.Equal(t, tt.expected.L1Gas.MaxAmount, got.L1Gas.MaxAmount,
-				"L1Gas.MaxAmount mismatch")
-			assert.Equal(t, tt.expected.L1Gas.MaxPricePerUnit, got.L1Gas.MaxPricePerUnit,
-				"L1Gas.MaxPricePerUnit mismatch")
+				// Compare each field individually for better error messages
+				assert.Equal(t, tt.expected.L1Gas.MaxAmount, got.L1Gas.MaxAmount,
+					"L1Gas.MaxAmount mismatch")
+				assert.Equal(t, tt.expected.L1Gas.MaxPricePerUnit, got.L1Gas.MaxPricePerUnit,
+					"L1Gas.MaxPricePerUnit mismatch")
 
-			assert.Equal(t, tt.expected.L1DataGas.MaxAmount, got.L1DataGas.MaxAmount,
-				"L1DataGas.MaxAmount mismatch")
-			assert.Equal(t, tt.expected.L1DataGas.MaxPricePerUnit, got.L1DataGas.MaxPricePerUnit,
-				"L1DataGas.MaxPricePerUnit mismatch")
+				assert.Equal(t, tt.expected.L1DataGas.MaxAmount, got.L1DataGas.MaxAmount,
+					"L1DataGas.MaxAmount mismatch")
+				assert.Equal(
+					t,
+					tt.expected.L1DataGas.MaxPricePerUnit,
+					got.L1DataGas.MaxPricePerUnit,
+					"L1DataGas.MaxPricePerUnit mismatch",
+				)
 
-			assert.Equal(t, tt.expected.L2Gas.MaxAmount, got.L2Gas.MaxAmount,
-				"L2Gas.MaxAmount mismatch")
-			assert.Equal(t, tt.expected.L2Gas.MaxPricePerUnit, got.L2Gas.MaxPricePerUnit,
-				"L2Gas.MaxPricePerUnit mismatch")
-		})
-	}
+				assert.Equal(t, tt.expected.L2Gas.MaxAmount, got.L2Gas.MaxAmount,
+					"L2Gas.MaxAmount mismatch")
+				assert.Equal(t, tt.expected.L2Gas.MaxPricePerUnit, got.L2Gas.MaxPricePerUnit,
+					"L2Gas.MaxPricePerUnit mismatch")
+			})
+		}
+	})
+
+	maxUint128BigInt, ok := new(big.Int).SetString(maxUint128, 0)
+	require.True(t, ok)
+
+	// All previous tests + new test with custom fee limits.
+	tests = append(tests, []struct {
+		name          string
+		feeEstimation rpc.FeeEstimation
+		multiplier    float64
+		expected      rpc.ResourceBoundsMapping
+		feeLimit      FeeLimits
+	}{
+		{
+			name: "With fee limit + multiplier 1.5",
+			feeEstimation: rpc.FeeEstimation{
+				FeeEstimationCommon: rpc.FeeEstimationCommon{
+					L1GasPrice:        BigIntToFelt(big.NewInt(1_000_000)),
+					L1GasConsumed:     BigIntToFelt(big.NewInt(500_000)),
+					L1DataGasPrice:    BigIntToFelt(big.NewInt(1_000_000_000)),
+					L1DataGasConsumed: BigIntToFelt(big.NewInt(1_000_000)),
+					L2GasPrice:        BigIntToFelt(big.NewInt(800_000)),
+					L2GasConsumed:     BigIntToFelt(big.NewInt(200_000_000)),
+				},
+			},
+			feeLimit: FeeLimits{
+				L1GasPriceLimit:      rpc.U128("0x124f80"),    // 1_200_000
+				L1GasAmountLimit:     rpc.U64("0xf4240"),      // 1_000_000
+				L1DataGasPriceLimit:  rpc.U128("0xf4240"),     // 1_000_000
+				L1DataGasAmountLimit: rpc.U64("0xe8d4a51000"), // 1_000_000_000_000
+				L2GasPriceLimit:      rpc.U128("0x7a120"),     // 500_000
+				L2GasAmountLimit:     rpc.U64("0xe8d4a51000"), // 1_000_000_000_000
+			},
+			multiplier: 1.5,
+			expected: rpc.ResourceBoundsMapping{
+				L1Gas: rpc.ResourceBounds{
+					// 1_000_000 * 1.5 = 1_500_000_
+					MaxPricePerUnit: "0x124f80", // 1_200_000, capped by the limit
+					// 500_000 * 1.5 = 750_000_
+					MaxAmount: "0xb71b0", // 750_000, within the limit
+				},
+				L1DataGas: rpc.ResourceBounds{
+					// 1_000_000_000 * 1.5 = 1_500_000_000_
+					MaxPricePerUnit: "0xf4240", // 1_000_000, capped by the limit
+					// 1_000_000 * 1.5 = 1_500_000_
+					MaxAmount: "0x16e360", // 1_500_000, within the limit
+				},
+				L2Gas: rpc.ResourceBounds{
+					// 800_000 * 1.5 = 1_200_000_
+					MaxPricePerUnit: "0x7a120", // 500_000, capped by the limit
+					// 200_000_000 * 1.5 = 300_000_000_
+					MaxAmount: "0x11e1a300", // 300_000_000, within the limit
+				},
+			},
+		},
+		{
+			name: "overflows, with only one limit set",
+			feeEstimation: rpc.FeeEstimation{
+				FeeEstimationCommon: rpc.FeeEstimationCommon{
+					L1GasPrice:        BigIntToFelt(maxUint128BigInt),
+					L1GasConsumed:     BigIntToFelt(maxUint128BigInt),
+					L1DataGasPrice:    BigIntToFelt(maxUint128BigInt),
+					L1DataGasConsumed: BigIntToFelt(maxUint128BigInt),
+					L2GasPrice:        BigIntToFelt(maxUint128BigInt),
+					L2GasConsumed:     BigIntToFelt(maxUint128BigInt),
+				},
+			},
+			feeLimit: FeeLimits{
+				L1GasPriceLimit: rpc.U128("0xf4240"), // 1_000_000
+			},
+			multiplier: 100,
+			// All outputs but the L1Gas.MaxPricePerUnit should be the max U128 and U64 values.
+			expected: rpc.ResourceBoundsMapping{
+				L1Gas: rpc.ResourceBounds{
+					MaxPricePerUnit: rpc.U128("0xf4240"), // The same as the limit.
+					MaxAmount:       rpc.U64(maxUint64),
+				},
+				L1DataGas: rpc.ResourceBounds{
+					MaxPricePerUnit: rpc.U128(maxUint128),
+					MaxAmount:       rpc.U64(maxUint64),
+				},
+				L2Gas: rpc.ResourceBounds{
+					MaxPricePerUnit: rpc.U128(maxUint128),
+					MaxAmount:       rpc.U64(maxUint64),
+				},
+			},
+		},
+	}...)
+
+	t.Run("Test CustomFeeEstToResBoundsMap", func(t *testing.T) {
+		for _, tt := range tests {
+			t.Run(tt.name, func(t *testing.T) {
+				got := CustomFeeEstToResBoundsMap(tt.feeEstimation, tt.multiplier, &tt.feeLimit)
+
+				// Compare each field individually for better error messages
+				assert.Equal(t, tt.expected.L1Gas.MaxAmount, got.L1Gas.MaxAmount,
+					"L1Gas.MaxAmount mismatch")
+				assert.Equal(t, tt.expected.L1Gas.MaxPricePerUnit, got.L1Gas.MaxPricePerUnit,
+					"L1Gas.MaxPricePerUnit mismatch")
+
+				assert.Equal(t, tt.expected.L1DataGas.MaxAmount, got.L1DataGas.MaxAmount,
+					"L1DataGas.MaxAmount mismatch")
+				assert.Equal(
+					t,
+					tt.expected.L1DataGas.MaxPricePerUnit,
+					got.L1DataGas.MaxPricePerUnit,
+					"L1DataGas.MaxPricePerUnit mismatch",
+				)
+
+				assert.Equal(t, tt.expected.L2Gas.MaxAmount, got.L2Gas.MaxAmount,
+					"L2Gas.MaxAmount mismatch")
+				assert.Equal(t, tt.expected.L2Gas.MaxPricePerUnit, got.L2Gas.MaxPricePerUnit,
+					"L2Gas.MaxPricePerUnit mismatch")
+			})
+		}
+	})
 }
 
 // TestTxnOptions tests the ApplyOptions method of the TxnOptions struct,